pyproject.toml
[tool.poetry]
name = "cosmrunner"
version = "0.1.0"
description = "python space shooter game"
authors = ["NEOZEO64"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pygame = "^2.5.2"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


